+++ src/global/vars.c
@@ -1,0 +1,5 @@
+#include "global/vars.h"
+
+#ifndef MESON_BUILD
+const char *g_TR2XVersion = "TR2X (non-Docker build)";
+#endif