+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 ## [Unreleased](https://github.com/rr-/TR2X/compare/stable...develop) - ××××-××-××
+- added version string to the inventory
 - fixed CDAudio not playing on certain versions (uses PaulD patch)
 - fixed TGA screenshots crashing the game